[fix]Refactored folder structure
