+++ TradePlay/src/components/card/small-video-card.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
-import { RemoveFromPlaylist } from "./add-to-playlist";
-import { RemoveFromHistory } from "./remove-from-history";
+import { RemoveFromPlaylist } from "./components/add-to-playlist";
+import { RemoveFromHistory } from "./components/remove-from-history";
 import { useVideos } from "../../context";
 
 const SmallVideoCard = ({ videoData, playlistId, remove, setPlaylistData }) => {
