+++ TradePlay/src/context/video-context/video-reducer.jsx
@@ -1,12 +1,13 @@
 import { changeCategory } from "../../utils";
+import { REDUCER_ACTION } from "../../utils/constants";
 
 const videoReducer = (state, action) => {
   switch (action.type) {
-    case "VIDEOS_DATA":
+    case REDUCER_ACTION.VIDEOS_DATA:
       return { ...state, videos: action.value };
-    case "CATEGORY_CHANGE":
+    case REDUCER_ACTION.CATEGORY_CHANGE:
       return changeCategory(state, action.value);
-    case "CLEAR_FILTER":
+    case REDUCER_ACTION.CLEAR_FILTER:
       return changeCategory(state, "");
   }
 };
