+++ TradePlay/src/pages/liked-video-page/liked-video-page.jsx
@@ -1,4 +1,3 @@
-// import { VideoCard } from "../../card/video-card";
 import { changeTitle } from "../../utils";
 import { useEffect } from "react";
 import { useAuth } from "../../context";
