+++ TradePlay/src/pages/search-result-page/search-result-page.jsx
@@ -1,6 +1,5 @@
 import { useParams } from "react-router-dom";
 import { useVideos } from "../../context/video-context/videos-context";
-// import { VideoCard } from "../../card/video-card";
 import { changeTitle } from "../../utils";
 import { useEffect } from "react";
 import { VideoCard } from "../../components/card/video-card";
