--- TradePlay/src/utils/add-to-history.jsx
@@ -1,5 +1,0 @@
-const addToHistory = (_id, videoDispatch) => {
-  videoDispatch({ type: "ADD_TO_HISTORY", value: _id });
-};
-
-export { addToHistory };
