+++ TradePlay/src/utils/constants.jsx
@@ -1,3 +1,13 @@
-const API_URL = "/api/";
+export const API_URL = "/api/";
 
-export { API_URL };
+export const REDUCER_ACTION = {
+  SET_TOKEN: "SET_TOKEN",
+  UPDATE_HISTORY: "UPDATE_HISTORY",
+  UPDATE_LIKED_VIDEOS: "UPDATE_LIKED_VIDEOS",
+  UPDATE_PLAYLISTS: "UPDATE_PLAYLISTS",
+  UPDATE_WATCHLATER: "UPDATE_WATCHLATER",
+  VIDEOS_DATA: "VIDEOS_DATA",
+  CATEGORY_CHANGE: "CATEGORY_CHANGE",
+  CLEAR_FILTER: "CLEAR_FILTER",
+};
+
