+++ TradePlay/src/utils/index.jsx
@@ -2,4 +2,3 @@ export { changeTitle } from "./change-title";
 export { authHeader } from "./auth-header";
 export { inList } from "./in-list";
 export { saveApiDataToContext, changeCategory } from "./video-context-methods";
-export { API_URL } from "./constants";
